requirements.txt
streamlit==1.29.0
imageio==2.31.1
imageio-ffmpeg==0.4.9
faster-whisper==1.0.1
openai==1.40.0
yt-dlp>=2024.10.14
ffmpeg-python==0.2.0
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
pyrebase4
firebase-admin